requirements.txt
flask
numpy
torch
transformers
spacy
pandas
dill
google-generativeai
NewsSentiment
